+++ cmd/traffic/cmd/manager/service.go
@@ -381,6 +381,7 @@ func (m *Manager) UpdateIntercept(ctx context.Context, req *rpc.UpdateInterceptR
 					DisplayBanner: action.AddPreviewDomain.DisplayBanner,
 					InterceptSpec: intercept.Spec,
 					Host:          action.AddPreviewDomain.Ingress.L5Host,
+					InjectHeaders: action.AddPreviewDomain.InjectHeaders,
 				})
 				if err != nil {
 					err = errors.Wrap(err, "systema: create domain")